Split storage.py (step 2)
Due to git limitation splitting requires two commits to track history across
all the files. Previous and this commit should be treated as one.

Splitting `datacube.storage.storage` into

storage.py -> ./_reproject.py
           -> ./_rio.py
           -> ../drivers/netcdf/_write.py

- `reproject_and_fuse` is now importable from `datacube.storage` directly

- "Datasource" classes moved into `datacube.storage._rio`

- Moved netcdf writer code in to drivers.netcdf

   These are mostly functions dealing with netcdf construction, used by internal
   tools `stacker/fixer`, yet they were imported for everyone from
   `datacube.core`, sometimes causing import loops.
