+++ datacube/drivers/netcdf/__init__.py
@@ -1,0 +1,8 @@
+from ._write import write_dataset_to_netcdf, create_netcdf_storage_unit
+from . import writer as netcdf_writer
+
+__all__ = (
+    'create_netcdf_storage_unit',
+    'write_dataset_to_netcdf',
+    'netcdf_writer',
+)
