+++ datacube/drivers/netcdf/writer.py
@@ -14,7 +14,7 @@ from datacube.utils import geometry, data_resolution_and_offset
 
 # pylint: disable=ungrouped-imports
 try:
-    from datacube.storage.netcdf_safestrings import SafeStringsDataset as Dataset
+    from ._safestrings import SafeStringsDataset as Dataset
 except TypeError:  # The above fails when netCDF4.Dataset is mocked, eg in RTD
     from netCDF4 import Dataset  # type: ignore
