+++ datacube_apps/stacker/stacker.py
@@ -22,8 +22,7 @@ import datacube
 from datacube.model import Dataset
 from datacube.model.utils import xr_apply, datasets_to_doc
 from datacube.utils import mk_part_uri
-from datacube.storage import netcdf_writer
-from datacube.storage.storage import create_netcdf_storage_unit
+from datacube.drivers.netcdf import create_netcdf_storage_unit, netcdf_writer
 from datacube.ui import task_app
 
 _LOG = logging.getLogger(__name__)
