+++ tests/test_driver.py
@@ -7,7 +7,7 @@ from datacube.drivers import new_datasource, reader_drivers, writer_drivers
 from datacube.drivers import index_drivers, index_driver_by_name
 from datacube.drivers.indexes import IndexDriverCache
 from datacube.storage import BandInfo
-from datacube.storage.storage import RasterDatasetDataSource
+from datacube.storage._rio import RasterDatasetDataSource
 from datacube.testutils import mk_sample_dataset
 from datacube.model import MetadataType
